+++ coroutines/src/commonMain/kotlin/dev/inmo/micro_utils/coroutines/FlowNullables.kt
@@ -1,0 +1,6 @@
+package dev.inmo.micro_utils.coroutines
+
+import kotlinx.coroutines.flow.*
+
+fun <T> Flow<T>.takeNotNull() = mapNotNull { it }
+fun <T> Flow<T>.filterNotNull() = takeNotNull()